added security on likes

+++ liquidvirtue/views.py
@@ -80,23 +80,24 @@ def get_next(request):
 	return render_to_response('get_next.html', {'lv_video_id': lv_video_id, 'title': title, 'youtube_video_id': youtube_video_id})
 
 def like(request, lv_video_id):
-	like_status = ''
-	if Like.objects.filter(user=request.user.id).filter(video=lv_video_id).exists():
-		v = Video.objects.get(pk=lv_video_id)
-		v.num_likes = v.num_likes - 1
-		if v.num_likes < 0:
-			v.num_likes = 0
-		v.save()
-		l = Like.objects.filter(user=request.user.id).filter(video=lv_video_id)
-		l.delete()
-		like_status = 'plus'
-	else:
-		v = Video.objects.get(pk=lv_video_id)
-		v.num_likes = v.num_likes + 1
-		v.save()
-		l = Like( video=v, user=UserProfile.objects.get(pk=request.user.id) )
-		l.save()
-		like_status = 'heart'
+	like_status = 'plus'
+	if request.user.is_authenticated():
+		if Like.objects.filter(user=request.user.id).filter(video=lv_video_id).exists():
+			v = Video.objects.get(pk=lv_video_id)
+			v.num_likes = v.num_likes - 1
+			if v.num_likes < 0:
+				v.num_likes = 0
+			v.save()
+			l = Like.objects.filter(user=request.user.id).filter(video=lv_video_id)
+			l.delete()
+			like_status = 'plus'
+		else:
+			v = Video.objects.get(pk=lv_video_id)
+			v.num_likes = v.num_likes + 1
+			v.save()
+			l = Like( video=v, user=UserProfile.objects.get(pk=request.user.id) )
+			l.save()
+			like_status = 'heart'
 		
 	return render_to_response('like.html', {'like_status': like_status})
